requirements.txt
pydantic>=2.0
pydantic-settings>=2.0
requests>=2.31
python-dotenv>=1.0